Create useAlert hook
